Initial user class implementation
Very simplistic still, does not do much other than fetch user id, custom greeting/title and last seen timestamp, but it's the first step towards migrating this with the admin system.

+++ src/game/etj_time_utilities.h
@@ -164,7 +164,7 @@ struct Time {
                         date.abbrevMonths[this->date.mon - 1], this->date.year);
   }
 
-  static Time fromInt(int input) {
+  static Time fromInt(int64_t input) {
     time_t value = input;
     tm *t = std::gmtime(&value);
